merge@2019-06-30

+++ app/jobs/analysis_job.rb
@@ -3,17 +3,30 @@ class AnalysisJob < ActiveJob::Base
 
   def perform(analysis_id)
     analysis = Analysis.find(analysis_id)
+    script_dir = Rails.root.join('scripts')
+    FileUtils.mkdir_p(File.join(script_dir, 'tmp'))
+
     args = [
       "'#{analysis.from.strftime('%F %T')}'",
       "'#{analysis.to.strftime('%F %T')}'",
+      analysis.pair,
       analysis.batch_size,
     ]
-    script_dir = File.join(Rails.root, 'scripts')
-    FileUtils.mkdir_p(File.join(script_dir, 'tmp'))
-    ret = system "sudo docker exec regulus python /opt/scripts/learn.py #{args.join(' ')}"
-    FileUtils.mv(File.join(script_dir, 'tmp'), File.join(Rails.root, "tmp/models/#{analysis_id}"))
-    analysis.update!(:state => 'completed')
-    AnalysisMailer.finished(analysis, ret).deliver_now
+    command = "sudo docker exec regulus python /opt/scripts/learn.py #{args.join(' ')}"
+    ret = system(command)
+    raise StandardError unless ret
+
+    from = File.join(script_dir, 'tmp')
+    to = Rails.root.join('tmp', 'models', analysis_id.to_s)
+    FileUtils.mv(from, to)
+    File.open(File.join(to, 'metadata.yml'), 'w') do |file|
+      YAML.dump({'pair' => analysis.pair}, file)
+    end
+    analysis.update!(state: 'completed')
+    AnalysisMailer.completed(analysis).deliver_now
     FileUtils.rm_rf("#{Rails.root}/tmp/models/#{analysis_id}")
+  rescue StandardError
+    analysis.update!(state: 'error')
+    AnalysisMailer.error(analysis).deliver_now
   end
 end
